hot fix

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -5,21 +5,21 @@
  *
  * Return: the square root of n, or -1 if n does not have a natural square root
  */
-int sqrt(int n, int i);
+double sqrt(int n, int i);
 int _sqrt_recursion(int n)
 {
-if (n < 0)
-{
-return (-1);
-}
-else if (n == 0 || n == 1)
-{
-return (n);
-}
-else
-{
-return (sqrt(n, 1));
-}
+    if (n < 0)
+    {
+        return (-1);
+    }
+    else if (n == 0 || n == 1)
+    {
+        return (n);
+    }
+    else
+    {
+        return (sqrt(n, 1));
+    }
 }
 /**
  * sqrt - helper function to recursively calculate square root
@@ -28,18 +28,18 @@ return (sqrt(n, 1));
  *
  * Return: the square root of n, or -1 if n does not have a natural square root
  */
-int sqrt(int n, int i)
+double sqrt(int n, int i)
 {
-if (i * i == n)
-{
-return (i);
-}
-else if (i * i > n)
-{
-return (-1);
-}
-else
-{
-return (sqrt(n, i + 1));
-}
+    if (i * i == n)
+    {
+        return (i);
+    }
+    else if (i * i > n)
+    {
+        return (-1);
+    }
+    else
+    {
+        return (sqrt(n, i + 1));
+    }
 }
